requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
langchain-community>=0.0.10
openai
faiss-cpu
pdfminer.six>=20221105
unstructured
pi-heif
langchain-openai
pdfplumber